requirements.txt
python-dotenv
requests
psycopg2-binary
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
hubspot-api-client
mailchimp-marketing
asana
slack-sdk
pytest
